requirements.txt
Flask
pandas
mysql-connector-python
waitress
